llm/requirements.txt
google-generativeai==0.7.2
python-dotenv==1.0.0
fastapi==0.115.2
uvicorn==0.32.0
